pyproject.toml
[tool.poetry]
name = "swali-ai"
version = "0.1.0"
description = "RAG-powered interview preparation system for LeetCode, system design, and technical interviews"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "app", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"

# FastAPI & Server
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"

# RAG Components
chromadb = "^0.6.0"
sentence-transformers = "^2.3.1"
anthropic = "^0.18.1"

# Data Processing
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
numpy = "<2.0"
google-generativeai = "^0.8.6"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"

# Code Quality
ruff = "^0.1.11"
mypy = "^1.8.0"

# Evaluation & Experiments
pandas = "^2.1.4"
matplotlib = "^3.8.2"
seaborn = "^0.13.1"

[tool.poetry.group.notebooks.dependencies]
# For exploratory work
jupyter = "^1.0.0"
ipywidgets = "^8.1.1"

[tool.poetry.scripts]
# CLI commands for convenience
collect-data = "scripts.run_collection:main"
start-api = "app.main:start"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["backend"]
addopts = "-v --cov=app --cov-report=html"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
